requirements.txt
# Core
pywebview>=5.0
llama-index>=0.14.0
llama-parse>=0.6.0
chromadb>=0.5.0

# LlamaIndex integrations
llama-index-vector-stores-chroma>=0.5.0
llama-index-llms-anthropic>=0.11.0
llama-index-llms-gemini>=0.6.0
llama-index-embeddings-openai>=0.6.0

# Document Processing
liteparse>=1.0.0
opendataloader-pdf>=2.0.0
yt-dlp>=2024.0.0
youtube-transcript-api>=0.6.0

# LLM Providers
anthropic>=0.40.0
openai>=1.50.0
google-generativeai>=0.8.0

# Storage
sqlalchemy>=2.0.0
keyring>=25.0.0

# Utilities
tiktoken>=0.7.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
